Minor fixes for command manager
